Fix firefox angles in P+ mode
Firefox angles are broken such that they do not work if you are holding the B button. I copied the code beneath it for what should be the correct firefox angles and placed them under the `(bs.b && (bs.mx != bs.my))` condition. I also labeled it in the code. The only change here should be that firefox angles work regardless of whether B is pressed or not.