Fixes for shield tilt using modifiers & LS => Taunt toggle
-- Corrects angles when using MX/MY to allow for shield tilt in all directions without roll or spotdodge --
Coords for MX/MY + L/R were `(0.82, 0.35)` and `(0.51, 0.82)` respectively. In P+, the threshold for roll/spotdodge is 0.73 in any direction. I have changed the values to `(0.72, 0.31)` and `(0.44, 0.72)`, which allows for shield tilt in any direction without accidental rolls while retaining the original angles.

-- LS => Taunt toggle --
I also added a toggle in code similar to true Z press labeled as `LSisDTaunt`. This is set as `true` by default and changes the LS button to DTaunt. In P+, there is no light shield, so light press trigger has extremely limited and niche utility, where its only unique function would be to be a dedicated L-cancel button to avoid tech lockout (Z does not do this in P+). Since this is already handled with the `trueZPress` option, it makes more sense to have LS be a taunt that can be remapped in-game to allow for more custom configurations. This can be set back to its original LS function by changing `LSisDTaunt` to `false`.